+++ test/td_ie_web/controllers/ingest_execution_controller_test.exs
@@ -7,6 +7,8 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
 
   import Mox
 
+  alias TdCore.Search.MockIndexWorker
+
   @create_attrs %{
     end_timestamp: ~N[2010-04-17 14:00:00.000000],
     start_timestamp: ~N[2010-04-17 14:00:00.000000],
@@ -32,8 +34,9 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
 
   setup do
     start_supervised!(TdIe.Cache.IngestLoader)
-    start_supervised!(TdIe.Search.IndexWorker)
-    start_supervised!(TdIe.Search.Cluster)
+    start_supervised!(TdCore.Search.Cluster)
+    start_supervised!(TdCore.Search.IndexWorker)
+
     :ok
   end
 
@@ -50,8 +53,6 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
   describe "create ingest_execution" do
     @tag authentication: [role: "admin"]
     test "renders ingest_execution when data is valid", %{conn: conn, swagger_schema: schema} do
-      SearchHelpers.expect_bulk_index()
-
       %{ingest_id: ingest_id} = insert(:ingest_version)
 
       assert %{"data" => %{"id" => id}} =
@@ -79,6 +80,8 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
                "description" => "some description",
                "records" => 10
              }
+
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     @tag authentication: [role: "admin"]
@@ -102,8 +105,6 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
       conn: conn,
       swagger_schema: schema
     } do
-      SearchHelpers.expect_bulk_index()
-
       insert(:ingest_version, name: "nombre sobrescrito")
 
       assert %{"data" => _data} =
@@ -114,6 +115,8 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
                )
                |> validate_resp_schema(schema, "IngestExecutionByNameResponse")
                |> json_response(:created)
+
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     @tag authentication: [role: "admin"]
@@ -150,8 +153,6 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
   describe "update ingest_execution" do
     @tag authentication: [role: "admin"]
     test "renders ingest_execution when data is valid", %{conn: conn, swagger_schema: schema} do
-      SearchHelpers.expect_bulk_index()
-
       %{ingest_id: ingest_id} = insert(:ingest_version)
       %{id: id} = ingest_execution = insert(:ingest_execution, ingest_id: ingest_id)
 
@@ -183,6 +184,8 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
                "description" => "some updated description",
                "records" => 11
              }
+
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     @tag authentication: [role: "admin"]
@@ -204,8 +207,6 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
   describe "delete ingest_execution" do
     @tag authentication: [role: "admin"]
     test "deletes chosen ingest_execution", %{conn: conn} do
-      SearchHelpers.expect_bulk_index()
-
       %{ingest_id: ingest_id} = insert(:ingest_version)
       ingest_execution = insert(:ingest_execution, ingest_id: ingest_id)
 
@@ -218,6 +219,8 @@ defmodule TdIeWeb.IngestExecutionControllerTest do
       assert_error_sent(:not_found, fn ->
         get(conn, Routes.ingest_ingest_execution_path(conn, :show, ingest_id, ingest_execution))
       end)
+
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
   end
 end