+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-6234] Use td-core-lib for elastic search and reindex
+
 ## [5.18.0] 2023-11-13
 
 ### Changed